requirements.txt
gpxpy==1.6.1
Pillow==10.1.0
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
pytz==2023.3.post1
shiboken6==6.6.0
